+++ examples/electron/test/render.test.js
@@ -7,10 +7,10 @@ RxDB.plugin(require('pouchdb-adapter-idb'));
  * this tests run inside of the browser-windows so we can ensure
  * rxdb works correctly
  */
-module.exports = (function() {
-    const runTests = async function() {
+module.exports = (function () {
+    const runTests = async function () {
         // issue #587 Icorrect working attachments in electron-render
-        await (async function() {
+        await (async function () {
             const db = await RxDB.create({
                 name: 'foobar587' + new Date().getTime(),
                 adapter: 'idb',
@@ -44,15 +44,21 @@ module.exports = (function() {
             });
             assert.ok(doc);
 
+            const attachmentData = 'foo bar asldfkjalkdsfj';
             const attachment = await doc.putAttachment({
                 id: 'cat.jpg',
-                data: 'foo bar asldfkjalkdsfj',
+                data: attachmentData,
                 type: 'text/plain'
             });
             assert.ok(attachment);
 
+
+            // issue #1371 Attachments not working in electron renderer with idb
+            const readData = await attachment.getStringData();
+            assert.equal(readData, attachmentData);
+
             await db.destroy();
-        })();
+        }());
     };
     return runTests;
-})();
+}());