+++ src/plugins/attachments.ts
@@ -79,6 +79,18 @@ export const blobBufferUtil = {
                 const text = e.target.result;
                 res(text);
             });
+
+            const blobBufferType = Object.prototype.toString.call(blobBuffer);
+
+            /**
+             * in the electron-renderer we have a typed array insteaf of a blob
+             * so we have to transform it.
+             * @link https://github.com/pubkey/rxdb/issues/1371
+             */
+            if (blobBufferType === '[object Uint8Array]') {
+                blobBuffer = new Blob([blobBuffer]);
+            }
+
             reader.readAsText(blobBuffer);
         });
     }