+++ test/helper/schemas.js
@@ -760,3 +760,29 @@ export const humanMinimalBroken = {
     },
     required: ['passportId', 'broken']
 };
+
+
+/**
+ * used in the graphql-test
+ * contains timestamp
+ */
+export const humanWithTimestamp = {
+    version: 0,
+    type: 'object',
+    properties: {
+        id: {
+            type: 'string',
+            primary: true
+        },
+        name: {
+            type: 'string'
+        },
+        age: {
+            type: 'number'
+        },
+        updatedAt: {
+            type: 'number'
+        }
+    },
+    required: ['id', 'name', 'age', 'updatedAt']
+};