+++ test/unit/server.test.js
@@ -58,7 +58,7 @@ config.parallel('server.test.js', () => {
         clientCollection.database.destroy();
         serverCollection.database.destroy();
     });
-    it('should send cors when defined', async function () {
+    it('should send cors when defined for missing origin', async function () {
         this.timeout(12 * 1000);
         const port = nexPort();
         const serverCollection = await humansCollection.create(0);
@@ -69,20 +69,69 @@ config.parallel('server.test.js', () => {
         });
         const colUrl = 'http://localhost:' + port + '/db/human';
 
-
-        const corsKey = 'Access-Control-Allow-Origin'.toLowerCase();
         await new Promise((res, rej) => {
             requestR({
                 method: 'GET',
-                url: colUrl
+                url: colUrl,
             }, (error, response) => {
                 if (error) rej(error);
-                const found = Object.entries(response.headers)
-                    .find(([k, v]) => {
-                        if (k.toLowerCase() === corsKey && v === '*') return true;
-                        else return false;
-                    });
-                if (!found) {
+
+                const responseHeaders = Object.keys(response.headers).reduce((acc, header) => ({
+                    ...acc,
+                    [header.toLowerCase()]: response.headers[header]
+                }), {});
+                const originHeaderName = 'Access-Control-Allow-Origin'.toLowerCase();
+                const credentialsHeaderName = 'Access-Control-Allow-Credentials'.toLowerCase();
+
+                const hasOriginHeader = responseHeaders[originHeaderName] === '*';
+                const hasCredentialsHeader = responseHeaders[credentialsHeaderName] === 'true';
+
+                if (!hasOriginHeader || !hasCredentialsHeader) {
+                    rej(
+                        new Error(
+                            'cors headers not set: ' +
+                            JSON.stringify(response.headers, null, 2)
+                        )
+                    );
+                } else res();
+            });
+        });
+
+        serverCollection.database.destroy();
+    });
+    it('should send cors when defined for present origin', async function () {
+        this.timeout(12 * 1000);
+        const port = nexPort();
+        const serverCollection = await humansCollection.create(0);
+        await serverCollection.database.server({
+            path: '/db',
+            port,
+            cors: true
+        });
+        const colUrl = 'http://localhost:' + port + '/db/human';
+
+        const origin = 'example.com';
+        await new Promise((res, rej) => {
+            requestR({
+                method: 'GET',
+                url: colUrl,
+                headers: {
+                    'Origin': origin,
+                }
+            }, (error, response) => {
+                if (error) rej(error);
+
+                const responseHeaders = Object.keys(response.headers).reduce((acc, header) => ({
+                    ...acc,
+                    [header.toLowerCase()]: response.headers[header]
+                }), {});
+                const originHeaderName = 'Access-Control-Allow-Origin'.toLowerCase();
+                const credentialsHeaderName = 'Access-Control-Allow-Credentials'.toLowerCase();
+
+                const hasOriginHeader = responseHeaders[originHeaderName] === origin;
+                const hasCredentialsHeader = responseHeaders[credentialsHeaderName] === 'true';
+
+                if (!hasOriginHeader || !hasCredentialsHeader) {
                     rej(
                         new Error(
                             'cors headers not set: ' +
@@ -229,5 +278,60 @@ config.parallel('server.test.js', () => {
         );
 
         db1.destroy();
+    });
+    describe('issues', () => {
+        describe('#1447 server path not working', () => {
+            it('use the path when given', async function () {
+                this.timeout(12 * 1000);
+                const port = nexPort();
+                const path = '/db2';
+                const serverCollection = await humansCollection.create(0);
+                await serverCollection.database.server({
+                    path,
+                    port
+                });
+
+                const colUrl = 'http://localhost:' + port + path + '/human';
+                const gotJson = await request(colUrl);
+                const got = JSON.parse(gotJson);
+                assert.equal(got.doc_count, 1);
+
+                serverCollection.database.destroy();
+            });
+            it('use the path with ending slash', async function () {
+                this.timeout(12 * 1000);
+                const port = nexPort();
+                const path = '/db3/';
+                const serverCollection = await humansCollection.create(0);
+                await serverCollection.database.server({
+                    path,
+                    port
+                });
+
+                const colUrl = 'http://localhost:' + port + path + 'human';
+                const gotJson = await request(colUrl);
+                const got = JSON.parse(gotJson);
+                assert.equal(got.doc_count, 1);
+
+                serverCollection.database.destroy();
+            });
+            it('should be able to use the root /', async function () {
+                this.timeout(12 * 1000);
+                const port = nexPort();
+                const path = '/';
+                const serverCollection = await humansCollection.create(0);
+                await serverCollection.database.server({
+                    path,
+                    port
+                });
+
+                const colUrl = 'http://localhost:' + port + path + 'human';
+                const gotJson = await request(colUrl);
+                const got = JSON.parse(gotJson);
+                assert.equal(got.doc_count, 1);
+
+                serverCollection.database.destroy();
+            });
+        });
     });
 });